frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["JoaquinAA <joaquin.alarcon@correounivalle.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
#rembg = "2.0.50"
Pillow = "9.3.0"
streamlit = "1.21.0"
numba = "0.57"
numpy = "1.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"